requirements.txt
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
elevenlabs==1.3.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fsspec==2024.6.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiter==0.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
natsort==8.4.0
numpy==2.0.0
openai==1.35.10
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
selenium==4.22.0
setuptools==70.2.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
tokenizers==0.19.1
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
undetected-chromedriver==3.5.5
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
